+++ checkpoint/orbax/__init__.py
@@ -71,4 +71,4 @@ except RuntimeError:
 
 # A new PyPI release will be pushed everytime `__version__` is increased.
 # Also modify version and date in CHANGELOG.
-__version__ = '0.5.9'
+__version__ = '0.5.10'