+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.10] - 2024-04-22
+
 ### Added
 - Add path package to export symbols, also add Step rst docs.
 - Create composite step `NameFormat`.